(#106)refactor: 응답 DTO 파일 위치 수정했습니다.
